Check if video width/height meets the video encoder requirements
This will avoid an encoding error when selecting with -s/--size
option a width/height value that does not meet the video encoder
requirement.

When selecting the recording region with mouse (-S/--select-region
option) the video width/height will be automatically changed to the
immediately higher number that meets the video encoder requirement.