Force width and height to be a multiple of 8 (to all video encoders)
This is a requirement for x265, kvazaar, theora, vp8 and vp9 video
encoders and was already being forced for them.

For x264, h264_nvenc and hevc_nvenc video encoders this is actually
not required, but it will avoid speedloss with them. Otherwise, the
following ffmpeg warning can appear, specially when using text
watermark:
'[swscaler @ 0x556f65c058a0] Warning: data is not aligned! This can
lead to a speedloss'