check_cmd_line(): check only values entered by command line
The default values do not need to be checked because they are
already correct.

Previous behavior was to check/validate the values of all options
even if they were not entered at command line by the user. That
behavior allowed to check values if they have been entered directly
at the source code and was mainly for debugging purposes. The user
is not supposed to change the source code, so this previous
behavior can be changed.

Fixes #2