Do not allow to use floats for video framerate when live streaming
When live streaming, the video framerate (fps) cannot be a
floating point number because it will be multiplied by two to
obtain the gop value. This multiplication will be done using the
shell simple aritmetic operator, which does not support floats.
This will make bc to be required only when using fade effect.