Change icon for error and finish notifications
For the finish notification, screencast will try to use an icon
from the Oxygen icon theme. If it is not installed on system,
screencast will fallback to use an icon name from the standard
Freedesktop icon specification. The reason for this is the fact
that the Freedesktop icon specification does not define an icon
name that can reflect a finish/complete/success message.