Fall back to PulseAudio if ALSA is not detected in FFmpeg
Surprisingly, some FFmpeg builds do not have support for the ALSA
backend (like the Fedora build from the RPM Fusion repository).
In this case, screencast will try to fallback to PulseAudio
backend. If this fallback can be done, a warning message will be
displayed and the '-i' option can be used to specify a PulseAudio
input device name.