Change the default audio input to value 'default'
The default audio input was changed to the value 'default'
(the default ALSA input device).

The recently released PulseAudio 11.0 broke something in FFmpeg?
Independently of this, defaulting to the value 'default' seems to
be more suitable for everyone, and also seems more logical since
the default audio backend used by screncast is ALSA.