Add checks for the needed FFmpeg decoders, muxers and demuxers
This follows the approach to check for all the needed ffmpeg
components to avoid surprises.

Now all the ffmpeg component checks are made in the single
check_component() function. This increases simplicity and
eliminates the need for multiple ffmpeg check functions.

The functions check_ffmpeg_format() and check_ffmpeg_encoder()
were removed.