Improve checks for invalid combination of format and encoders
Check for a valid combination of container format and audio/video
encoders  before checking for component support in ffmpeg. This
will stress less the user in the case that he selects a format or
encoder that is not supported by the detected ffmpeg build and is
also part of invalid combination for format/encoders.