Make it easier to add (or remove) formats/encoders/others
This commit is some kind of an important rewrite of some parts,
specially the initial variables declaration and check_cmd_line().

Now, the middle of the code does not make any direct reference to
the formats, encoders, or other key parts, like the effects.
Instead, the formats/encoders/effects/others are setted only at
the first parts of the script. This increases code abstraction and
allows to add new formats/encoders/others by modifying just the
first parts of the script, and the rest of the code will still
handle the addition. For example, for adding support for a new
audio encoder, just create the corresponding settings function
on the 'audio encoder settings' section and add it the the nearby
variable $supported_audiocodecs_all, without the need to mess with
the rest of the code.

This should have already been done previosly, but while the code
was not so big this was not much needed. It became mandatory as
the code increased in complexity and size by the additions of new
features and components.

At a first glance, it looks like that things are more complex now,
but in fact it is not. These changes will make it much more easier
to to code maintenance and add/remove formats/encoders/others.

Some other changes are also on this commit, since many thing were
touched. But they are too much to document here. Mostly small
things.

Probably something will break with this commit. I hope not. Fixes
for any regressions will come if something can be found.