Change streaming and one step settings to minimize audio problems
Recording in a one step process (record + encode at the same time)
can cause audio problems, like cut-offs and desyncs. This commit
tries to minimize these problems. Note that the problematic thing
in this case is not the audio settings, but the video ones. Faster
presets tends to minimize this problems, specially 'ultrafast'.

Also updated to documentation to reinforce these problems, not
recommending the use of --one-step/-1 with software-based encoders.