tests: add checks for some required ffmpeg components
Tests are performed with mp4 and flv container formats, and also
with aac and x264 encoders. Exit with error before starting the
tests if these requirements are not met.

This will prevent tests from running with ffmpeg builds that do not
met this requirement (for example, openSUSE), avoiding false failed
results in such platforms.