Refactor temporary directory creation and temporary video saving
The directory for temporary files now defaults only to
${XDG_CACHE_HOME}/screencast. No more ${HOME}/.screencast
is used. It also does not create a tmp.XXXXXX directory
inside the base temporary directory anymore. If the user
wants to use another directory for temporary files he can
use the -t option as before.

Note: if $XDG_CACHE_HOME is not setted, it will default
to ${HOME}/.cache as defined by the XDG specification,
meaning that the screencast directory for temporary
files will be at ${HOME}/.cache/screencast.

When the user wants to keep the temporary/lossless video
by using the -K option, it will be now moved to the
output directory. This creates a better consistency in
the location of the final saved files.